Splitting up `AggregationWizard` in multiple components.
